requirements.txt
numpy==1.26.4
opencv-contrib-python
fastapi==0.110.1
firebase-admin==6.5.0
gunicorn==20.1.0
uvicorn==0.14.0
Pillow
torch
torchvision
tqdm
python-multipart